Move normalized_raster methods from util to gdal
