+++ src/layman/layer/filesystem/__init__.py
@@ -1,6 +1,6 @@
 from layman.common import empty_method
 from layman.common.filesystem import util as common_util
-from . import util
+from . import util, gdal
 
 check_username = empty_method
 check_new_layername = empty_method
@@ -13,7 +13,7 @@ ensure_whole_user = common_util.ensure_whole_user
 
 def delete_workspace(workspace):
     common_util.delete_workspace(workspace)
-    util.delete_normalized_raster_workspace(workspace)
+    gdal.delete_normalized_raster_workspace(workspace)
 
 
 delete_whole_user = delete_workspace
