+++ src/layman/layer/filesystem/gdal.py
@@ -2,15 +2,15 @@ import os
 import shutil
 import subprocess
 from osgeo import gdal
-from layman import patch_mode
+from layman import patch_mode, settings
 from layman.common import empty_method, empty_method_returns_dict
-from . import input_file, util
+from . import input_file
 
 PATCH_MODE = patch_mode.DELETE_IF_DEPENDANT
 
 
 def get_layer_info(workspace, layer):
-    gdal_path = util.get_normalized_raster_layer_main_filepath(workspace, layer)
+    gdal_path = get_normalized_raster_layer_main_filepath(workspace, layer)
     if os.path.exists(gdal_path):
         return {
             'layername': layer,
@@ -33,7 +33,7 @@ patch_layer = empty_method
 
 def delete_layer(username, layername):
     try:
-        shutil.rmtree(util.get_normalized_raster_layer_dir(username, layername))
+        shutil.rmtree(get_normalized_raster_layer_dir(username, layername))
     except FileNotFoundError:
         pass
 
@@ -51,7 +51,7 @@ def get_color_interpretations(filepath):
 def normalize_raster_file_async(workspace, layer, input_path, crs_id):
     color_interp = get_color_interpretations(input_path)
     assert color_interp == ['Red', 'Green', 'Blue']
-    result_path = util.get_normalized_raster_layer_main_filepath(workspace, layer)
+    result_path = get_normalized_raster_layer_main_filepath(workspace, layer)
     bash_args = [
         'gdalwarp',
         '-of', 'GTiff',
@@ -75,3 +75,27 @@ def normalize_raster_file_async(workspace, layer, input_path, crs_id):
     process = subprocess.Popen(bash_args, stdout=subprocess.PIPE,
                                stderr=subprocess.STDOUT)
     return process
+
+
+def get_normalized_raster_workspace_dir(workspace):
+    return os.path.join(settings.LAYMAN_NORMALIZED_RASTER_DATA_DIR, 'workspaces', workspace)
+
+
+def get_normalized_raster_layer_dir(workspace, layer):
+    return os.path.join(get_normalized_raster_workspace_dir(workspace), 'layers', layer)
+
+
+def get_normalized_raster_layer_main_filepath(workspace, layer):
+    return os.path.join(get_normalized_raster_layer_dir(workspace, layer), f"{layer}.tif")
+
+
+def ensure_normalized_raster_layer_dir(workspace, layer):
+    gdal_dir = get_normalized_raster_layer_dir(workspace, layer)
+    os.makedirs(gdal_dir, exist_ok=True)
+
+
+def delete_normalized_raster_workspace(workspace):
+    try:
+        os.rmdir(get_normalized_raster_workspace_dir(workspace))
+    except FileNotFoundError:
+        pass
